refactor: move github-client into self-titled fiel
